Make Changelog class eagerly load and parse file.
Rather than reading and parsing the file on each access,
the class now eagerly reads the file in the constructor,
and immediately parses it into sections.

Since the section parsing is incredibly simple, the overhead
of splitting into sections should be minimal compared to loading
the file from disk.
Also, the existing code read the entire file, even if it only
needs the newest entry of the changelog, so parsing the entire
file once and for all should not be significantly slower,
and should be much faster on repeated access.

There should be no reason to create a Changelog and not query
it at all.